backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
mysql-connector-python==8.2.0
boto3==1.29.0
python-dotenv==1.0.0
Pillow==10.1.0
google-cloud-aiplatform
supabase>=2.27.0
websockets==15.0
python-dateutil>=2.8.2
PyJWT>=2.8.0
bcrypt>=4.1.0
Werkzeug>=3.0.0
